+++ functions/db.php
@@ -141,3 +141,11 @@ function isset_email($connection, $email){
     }
     return null;
 }
+
+function check_user($connection, $form){
+    $email = mysqli_real_escape_string($connection, $form['email']);
+    $sql = "SELECT * FROM users WHERE email = '$email'";
+    $res = mysqli_query($connection, $sql);
+    $user = $res ? mysqli_fetch_array($res, MYSQLI_ASSOC) : null;
+    return $user;
+}
