+++ lot.php
@@ -1,6 +1,6 @@
 <?php
 date_default_timezone_set("Europe/Moscow");
-
+session_start();
 require_once ('functions/db.php');
 require_once ('functions/lot_validate.php');
 require_once ('functions/template.php');
