+++ templates/login.php
@@ -8,18 +8,18 @@
             <?php endforeach; ?>
         </ul>
     </nav>
-    <form class="form container <?php if($errors): ?>form--invalid<?php endif;?>""  action="https://echo.htmlacademy.ru" method="post"> <!-- form--invalid -->
-    <h2>Вход</h2>
-    <div class="form__item <?php if($errors): ?>form__item--invalid<?php endif;?>"> <!-- form__item--invalid -->
-        <label for="email">E-mail*</label>
-        <input id="email" type="text" name="email" placeholder="Введите e-mail" required>
-        <span class="form__error">Введите e-mail</span>
-    </div>
-    <div class="form__item form__item--last <?php if($errors): ?>form__item--invalid<?php endif;?>">
-        <label for="password">Пароль*</label>
-        <input id="password" type="password" name="password" placeholder="Введите пароль" required>
-        <span class="form__error">Введите пароль</span>
-    </div>
-    <button type="submit" class="button">Войти</button>
+    <form class="form container <?php if($errors): ?>form--invalid<?php endif;?>"  action="../login.php" enctype="application/x-www-form-urlencoded" method="post"> <!-- form--invalid -->
+        <h2>Вход</h2>
+        <div class="form__item <?php if($errors): ?>form__item--invalid<?php endif;?>"> <!-- form__item--invalid -->
+            <label for="email">E-mail*</label>
+            <input id="email" type="text" name="email" placeholder="Введите e-mail" value="<?= $form['email'] ?>" required>
+            <span class="form__error"><?= $errors['email']; ?></span>
+        </div>
+        <div class="form__item form__item--last <?php if($errors): ?>form__item--invalid<?php endif;?>">
+            <label for="password">Пароль*</label>
+            <input id="password" type="password" name="password" placeholder="Введите пароль" value="<?= $form['password'] ?>" required>
+            <span class="form__error"> <?= $errors['password']; ?> </span>
+        </div>
+        <button type="submit" class="button">Войти</button>
     </form>
 </main>
